Base api server -> Main

+++ API_SERVER/app/Http/Controllers/Api/ArmasController.php
@@ -6,10 +6,16 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreUpdateArmaRequest;
 use App\Http\Resources\ArmasResource;
 use App\Models\Arma;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 
 class ArmasController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('verifyTokenSuap')->only(['store', 'update', 'destroy']);
+    }
+
     public function index()
     {
         $armas = Arma::all();
@@ -22,6 +28,9 @@ class ArmasController extends Controller
         //     if (!Auth::check()) {
         //         return response()->json(['message' => 'Você precisa estar autenticado para executar essa função', 403]);
         //     }
+        if (empty($request->bearerToken())) {
+            return response()->json(['message' => 'Token de acesso faltando ou inválido', 'status' => 401], 401);
+        }
 
         $data = $request->validated();
 
@@ -35,7 +44,7 @@ class ArmasController extends Controller
         $armas = Arma::where('index', '=', $index)->first();
 
         if (!$armas) {
-            return response()->json(['message' => 'Index não corresponde a nenhum resultado', 'status'=>404]);
+            return response()->json(['message' => 'Index não corresponde a nenhum resultado', 'status' => 404], 404);
         }
 
         return new ArmasResource($armas);
@@ -47,10 +56,14 @@ class ArmasController extends Controller
         //     return response()->json(['message' => 'Você precisa estar autenticado para executar essa função', 403]);
         // }
 
+        if (empty($request->bearerToken())) {
+            return response()->json(['message' => 'O usuário precisa estar autenticado', 'status' => 401], 401);
+        }
+
         $armas = Arma::where('index', '=', $index)->first();
 
         if (!$armas) {
-            return response()->json(['message' => 'Index não corresponde a nenhum resultado', 'status'=>404]);
+            return response()->json(['message' => 'Index não corresponde a nenhum resultado', 'status' => 404], 404);
         }
 
         $data = $request->validated();
@@ -60,20 +73,24 @@ class ArmasController extends Controller
         return new ArmasResource($armas);
     }
 
-    public function destroy(string $index)
+    public function destroy(string $index, Request $request)
     {
         // if (!Auth::check()) {
         //     return response()->json(['message' => 'Você precisa estar autenticado para executar essa função', 403]);
         // }
 
+        if (empty($request->bearerToken())) {
+            return response()->json(['message' => 'O usuário precisa estar autenticado', 'status' => 401], 401);
+        }
+
         $armas = Arma::where('index', '=', $index)->first();
 
         if (!$armas) {
-            return response()->json(['message' => 'Index não corresponde a nenhum resultado', 'status'=>404]);
+            return response()->json(['message' => 'Index não corresponde a nenhum resultado', 'status' => 404], 404);
         }
 
         $armas->delete();
 
-        return response()->json(['message' => 'A arma foi deletada com sucesso', 'status'=>200]);
+        return response()->json(['message' => 'A arma foi deletada com sucesso', 'status' => 200], 200);
     }
 }
